RTEformElement abstract
