+++ projects/ui-framework/src/lib/form-elements/rich-text-editor/rte.component.scss
